feat: Remove react resize detector
Contributes to [#2744](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2744)

V7 of ReactResizeObserver can result in the error shown in [#2744](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2744) but the error appears to be a timing issue of some description.

The author fixed a similar looking issue in V8, but it persists in our storybook. In V9 there were more extensive changes that stopped our current usage working.

This PR removes ReactResizeObserver in favor of wrapping ResizeObserver (now standard) in a hook.

*What did you change?*
- Removed react-resize-detector
- Created useResizeObserver hook.
- Checked it functioned as a replacement for all useResizeDetector hook
*How did you test and verify your work?*
- Storybook changing the size of things and adding new elements
- Ran tests